+++ tests/Xdg.hs
@@ -3,7 +3,10 @@ module Xdg where
 #if MIN_VERSION_base(4, 7, 0)
 import qualified Data.List as List
 import System.Environment (setEnv, unsetEnv)
-import System.FilePath ((</>), searchPathSeparator)
+import System.FilePath (searchPathSeparator)
+#if !defined(mingw32_HOST_OS)
+import System.FilePath ((</>))
+#endif
 #endif
 #include "util.inl"
 
@@ -24,12 +27,21 @@ main _t = do
   T(expectEq) () (home </> ".config/mow") =<< getXdgDirectory XdgConfig "mow"
 #endif
 
+  -- unset variables, so env doesn't affect test running
+  unsetEnv "XDG_DATA_HOME"
+  unsetEnv "XDG_CONFIG_HOME"
+  unsetEnv "XDG_CACHE_HOME"
+  xdgData   <- getXdgDirectory XdgData   "ff"
+  xdgConfig <- getXdgDirectory XdgConfig "oo"
+  xdgCache  <- getXdgDirectory XdgCache  "rk"
+
+  -- non-absolute paths are ignored, and the fallback is used
   setEnv "XDG_DATA_HOME"   "ar"
   setEnv "XDG_CONFIG_HOME" "aw"
   setEnv "XDG_CACHE_HOME"  "ba"
-  T(expectEq) () ("ar" </> "ff") =<< getXdgDirectory XdgData   "ff"
-  T(expectEq) () ("aw" </> "oo") =<< getXdgDirectory XdgConfig "oo"
-  T(expectEq) () ("ba" </> "rk") =<< getXdgDirectory XdgCache  "rk"
+  T(expectEq) () xdgData   =<< getXdgDirectory XdgData   "ff"
+  T(expectEq) () xdgConfig =<< getXdgDirectory XdgConfig "oo"
+  T(expectEq) () xdgCache  =<< getXdgDirectory XdgCache  "rk"
 
   unsetEnv "XDG_CONFIG_DIRS"
   unsetEnv "XDG_DATA_DIRS"